+++ packages/@expo/config-plugins/src/ios/__tests__/PrivacyInfo-test.ts
@@ -1,0 +1,69 @@
+import { vol } from 'memfs';
+import path from 'path';
+
+import { setPrivacyInfo, PrivacyInfo } from '../PrivacyInfo';
+
+jest.mock('fs');
+
+jest.mock('../utils/Xcodeproj', () => ({
+  getProjectName: () => 'testproject',
+  addBuildSourceFileToGroup: jest.fn(),
+}));
+
+const projectRoot = '/testproject';
+
+const project = {
+  name: 'test',
+  slug: 'test',
+};
+
+const mockConfig = {
+  //fill in relevant data here
+  modResults: {
+    hasFile: () => false,
+  },
+  modRequest: {
+    projectRoot,
+    platformProjectRoot: path.join(projectRoot, 'ios'),
+    modName: 'test',
+    platform: 'ios' as const,
+    introspect: false,
+  },
+  modRawConfig: project,
+  ...project,
+};
+
+const privacyManifests: PrivacyInfo = {
+  NSPrivacyAccessedAPITypes: [
+    {
+      NSPrivacyAccessedAPIType: 'NSPrivacyAccessedAPICategoryTestCategory',
+      NSPrivacyAccessedAPITypeReasons: ['TEST.TEST'],
+    },
+  ],
+  NSPrivacyCollectedDataTypes: [],
+  NSPrivacyTracking: true,
+  NSPrivacyTrackingDomains: ['test.com'],
+};
+
+const filePath = 'ios/testproject/PrivacyInfo.xcprivacy';
+
+const originalFs = jest.requireActual('fs');
+
+describe('withPrivacyInfo', () => {
+  afterEach(() => vol.reset());
+  it('adds PrivacyInfo.xcprivacy file to the project and merges with existing file', async () => {
+    // mock the data in the PrivacyInfo.xcprivacy file using vol
+    vol.fromJSON(
+      {
+        'ios/testproject/PrivacyInfo.xcprivacy': originalFs.readFileSync(
+          path.join(__dirname, 'fixtures/PrivacyInfo.xcprivacy'),
+          'utf-8'
+        ),
+      },
+      projectRoot
+    );
+
+    setPrivacyInfo(mockConfig, privacyManifests);
+    expect(vol.readFileSync(path.join(projectRoot, filePath), 'utf-8')).toMatchSnapshot();
+  });
+});